Variable type fix and cosmetics

+++ src/feather-32u4.cpp
@@ -1,4 +1,4 @@
-#include "feather-32u4.h"
+#include <feather-32u4.h>
 
 void led_on(){
     pinMode(LED_BUILTIN, OUTPUT);
@@ -14,8 +14,8 @@ float battery_voltage(){
   return analogRead(PIN_VBAT)/1024.0*6.6;
 }
 
-void sleep(int sleep_ms) {
-    int sleepMS =0;
+void sleep(unsigned long sleep_ms) {
+    unsigned long sleepMS =0;
 
     while(sleepMS <sleep_ms) {
         sleepMS +=Watchdog.sleep(min(sleep_ms -sleepMS, WD_MAX));
